+++ src/storage/store/store_sync/opendal.rs
@@ -221,8 +221,7 @@ mod tests {
     #[cfg_attr(miri, ignore)]
     fn filesystem() -> Result<(), Box<dyn Error>> {
         let path = tempfile::TempDir::new()?;
-        let mut builder = opendal::services::Fs::default();
-        builder.root(&path.path().to_string_lossy());
+        let builder = opendal::services::Fs::default().root(&path.path().to_string_lossy());
         let op = Operator::new(builder)?.finish().blocking();
         let store = OpendalStore::new(op);
         super::super::test_util::store_write(&store)?;